requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.9.1
idna==3.4
munch==2.5.0
packaging==23.0
python-dateutil==2.8.2
pytz==2022.7.1
shapely==2.0.1
six==1.16.0
urllib3==1.26.14
minio
numpy
pandas
pyproj
geopandas
requests
pendulum
apache-airflow-providers-apache-spark
pyspark
sqlalchemy-trino